Bug 1526215 - Fix back button behavior in Import YAML dialog
Clicking back from the "Template Configuration" step can prevent the
user from being able to click next again in the wizard in the Import
YAML dialog. This is because the `ctrl.currentStep` value is not reset
to the first step title after the user clicks back. Sine the wizard is
watching for changes to `ctrl.currentStep` and the value never changes
when the user clicks next again, the wizard doesn't advance.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1526215